refactore code, rename
