+++ src/elements/MetricsSelector/metricsSelector.utils.js
@@ -18,7 +18,7 @@ under the Apache 2.0 license is conditioned upon your compliance with
 such restriction.
 */
 import { chain, isEmpty } from 'lodash'
-import { ML_RUN_INFRA } from '../../components/DetailsMetrics/detailsMetrics.utils'
+import { ML_RUN_INFRA } from '../../components/DetailsMetrics/detailsMetrics.util'
 
 export const metricsTypes = {
   metric: 'metric',
